:hammer: refactor: 폴더구조반영
